Version 0.1.39
疑似HTML(HLML:Html Like Markup Languageと勝手に命名)による記事投稿に対応。
特殊文字/複数段落投稿に対応。
トップページはサイト名のみ出力出力するようにした。
管理者ユーザは削除できないようにした。
ユーザIDと名前の空白は削除するようにした。
0.1.36以前のDB互換を想定して、プログラム側でも64KBチェックを施した。

+++ project/trunk/lib/state/scene/user/CSceneRemoveUser.php
@@ -60,6 +60,11 @@ class CSceneRemoveUser
 				{
 					throw new Exception(_('ログインしていないため受理不可。'));
 				}
+				$body =& $user->getEntity()->storage();
+				if($body['root'])
+				{
+					throw new Exception(_('管理者ユーザは削除不可。'));
+				}
 				if(!$user->delete())
 				{
 					throw new Exception(_('予期しない理由での削除失敗。'));
